+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# Unreleased
+- [New] New GetAllDocumentsOperation API.
+
 # 0.3.2 (2016-07-08)
 
 - [FIXED] Created explicit `Sort` and `TextIndexField` initalizers so they are exposed as public.